fixed a problem with new numpy version

+++ brainstorm/scorers.py
@@ -69,7 +69,7 @@ class Accuracy(Scorer):
     def __call__(self, true_labels, predicted, mask=None):
         if predicted.shape[1] > 1:
             predicted = predicted.argmax(1).reshape(-1, 1)
-        correct = predicted == true_labels
+        correct = (predicted == true_labels).astype(np.float)
         if mask is not None:
             correct *= mask
         return np.sum(correct)
@@ -82,7 +82,8 @@ class Hamming(Scorer):
         self.threshold = threshold
 
     def __call__(self, true_labels, predicted, mask=None):
-        correct = np.logical_xor(predicted < self.threshold, true_labels)
+        correct = np.logical_xor(predicted < self.threshold,
+                                 true_labels).astype(np.float)
         if mask is not None:
             correct *= mask
         return np.sum(correct) / true_labels.shape[1]
